Refactor ApplicantStatusChangeEmail
