+++ app/Modules/Visittransfer/Listeners/NotifyApplicantOfReferenceAcceptance.php
@@ -2,7 +2,7 @@
 
 namespace App\Modules\Visittransfer\Listeners;
 
-use App\Modules\Visittransfer\Notifications\ApplicantReferenceAccepted;
+use App\Modules\Visittransfer\Notifications\ApplicationReferenceAccepted;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use App\Modules\Visittransfer\Events\ReferenceAccepted;
 
@@ -16,6 +16,6 @@ class NotifyApplicantOfReferenceAcceptance implements ShouldQueue
     public function handle(ReferenceAccepted $event)
     {
         $reference = $event->reference;
-        $reference->application->account->notify(new ApplicantReferenceAccepted($reference));
+        $reference->application->account->notify(new ApplicationReferenceAccepted($reference));
     }
 }
