+++ app/Modules/Visittransfer/Listeners/NotifyApplicantOfReferenceCompletion.php
@@ -2,7 +2,7 @@
 
 namespace App\Modules\Visittransfer\Listeners;
 
-use App\Modules\Visittransfer\Notifications\ApplicantReferenceSubmitted;
+use App\Modules\Visittransfer\Notifications\ApplicationReferenceSubmitted;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use App\Modules\Visittransfer\Events\ReferenceUnderReview;
 
@@ -16,6 +16,6 @@ class NotifyApplicantOfReferenceCompletion implements ShouldQueue
     public function handle(ReferenceUnderReview $event)
     {
         $reference = $event->reference;
-        $reference->application->account->notify(new ApplicantReferenceSubmitted($reference));
+        $reference->application->account->notify(new ApplicationReferenceSubmitted($reference));
     }
 }
