+++ app/Modules/Visittransfer/Listeners/NotifyApplicantOfReferenceRejection.php
@@ -2,7 +2,7 @@
 
 namespace App\Modules\Visittransfer\Listeners;
 
-use App\Modules\Visittransfer\Notifications\ApplicantReferenceRejected;
+use App\Modules\Visittransfer\Notifications\ApplicationReferenceRejected;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use App\Modules\Visittransfer\Events\ReferenceRejected;
 
@@ -16,6 +16,6 @@ class NotifyApplicantOfReferenceRejection implements ShouldQueue
     public function handle(ReferenceRejected $event)
     {
         $reference = $event->reference;
-        $reference->application->account->notify(new ApplicantReferenceRejected($reference));
+        $reference->application->account->notify(new ApplicationReferenceRejected($reference));
     }
 }
